Added type init function
We have added an unfinished type init function. It initiates the
vtable. But it does not do anything to the super class vtable.
Furthermore this function is not called anywhere. We need to add
a initiation check variable or something (type_is_initiated) for
this reason. Anyway things are going on fine.